+++ backend/src/main/java/com/genesis/unipocket/expense/persistence/dto/ExpenseManualCreateArgs.java
@@ -2,6 +2,8 @@ package com.genesis.unipocket.expense.persistence.dto;
 
 import com.genesis.unipocket.expense.common.enums.Category;
 import com.genesis.unipocket.global.common.enums.CurrencyCode;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -17,8 +19,8 @@ public record ExpenseManualCreateArgs(
 		Category category,
 		String paymentMethod,
 		LocalDateTime occurredAt,
-		Double localAmount,
+		BigDecimal localAmount,
 		CurrencyCode localCurrency,
-		Double standardAmount,
+		BigDecimal standardAmount,
 		CurrencyCode standardCurrency,
 		String memo) {}
